Revert "Added preliminary firmware for modules"
