+++ platformio.ini
@@ -8,7 +8,7 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
-[env:bluepill]
+[env:blackpill_f401cc]
 platform = ststm32
-board = bluepill_f103c8
+board = blackpill_f401cc
 framework = arduino
