+++ src/main.cpp
@@ -1,12 +1,3 @@
-////////////////////////////////////////////////////////////////////////////////////////////////////
-//*FILE NAME:       main.cpp
-//*FILE DESC:       Main file for norahHome.
-//*FILE VERSION:    0.1.1
-//*FILE AUTHOR:     Favour Aina
-//*LAST MODIFIED:   Saturday, 1 April, 2023 20:19
-//*LICENSE:         Academic Free License
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
 #include <Arduino.h>
 
 void setup() {
